move CallbackRegistry to pro directory
